+++ include/podio/ROOTLegacyReader.h
@@ -91,7 +91,7 @@ public:
 private:
   std::pair<TTree*, unsigned> getLocalTreeAndEntry(const std::string& treename);
 
-  void createCollectionBranches(const std::vector<std::tuple<uint64_t, std::string, bool, unsigned int>>& collInfo);
+  void createCollectionBranches(const std::vector<std::tuple<uint32_t, std::string, bool, unsigned int>>& collInfo);
 
   podio::GenericParameters readEventMetaData();
